requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
attrs=21.4.0=pyhd8ed1ab_0
blas=1.0=mkl
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py38h578d9bd_1
colorama=0.4.4=pyh9f0ad1d_0
contextlib2=21.6.0=pyhd8ed1ab_0
cycler=0.11.0=pyhd8ed1ab_0
execnet=1.9.0=pyhd8ed1ab_0
freetype=2.10.4=h5ab3b9f_0
glob2=0.7=py_0
icu=67.1=he1b5a44_0
iniconfig=1.1.1=pyh9f0ad1d_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9b=habf39ab_1
kiwisolver=1.3.2=py38h295c915_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
lz4-c=1.9.2=heb0550a_3
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py38h5d868c9_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mock=4.0.3=py38h578d9bd_2
ncurses=6.2=he6710b0_1
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
olefile=0.46=py_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pyhd8ed1ab_0
path=16.4.0=py38h578d9bd_0
path.py=12.5.0=0
pillow=8.0.0=py38h9a89aac_0
pip=20.2.4=py38_0
pluggy=1.0.0=py38h578d9bd_2
py=1.11.0=pyh6c4a22f_0
pyparsing=3.0.7=pyhd8ed1ab_0
pytest=7.1.0=py38h578d9bd_0
pytest-shutil=1.7.0=pyhd8ed1ab_1
python=3.8.5=h7579374_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
readline=8.0=h7b6447c_0
setuptools=50.3.0=py38hb0f4dca_1
six=1.16.0=pyh6c4a22f_0
sqlite=3.33.0=h62c20be_0
termcolor=1.1.0=py_2
tk=8.6.10=hbc83047_0
tomli=2.0.1=pyhd8ed1ab_0
tornado=6.1=py38h25fe258_0
tqdm=4.63.0=pyhd8ed1ab_0
wheel=0.35.1=py_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
